Split up and move tbconstants.ts
