+++ extension/data/modules/modbar.js
@@ -5,6 +5,7 @@ import * as TBCore from '../tbcore.js';
 import * as TBHelpers from '../tbhelpers.js';
 import TBModule, {Module} from '../tbmodule.jsx';
 import * as TBui from '../tbui.js';
+import {icons} from '../util/icons.ts';
 import {purify} from '../util/purify.js';
 import {getSettingAsync, setSettingAsync} from '../util/settings.ts';
 
@@ -205,8 +206,8 @@ export default new Module({
     );
     const $modBar = $(`
 <div id="tb-bottombar">
-    <a class="tb-bottombar-hide tb-icons" href="javascript:void(0)">${TBui.icons.arrowLeft}</a>
-    <a class="tb-toolbar-new-settings tb-icons" href="javascript:void(0)" title="toolbox settings">${TBui.icons.settings}</a>
+    <a class="tb-bottombar-hide tb-icons" href="javascript:void(0)">${icons.arrowLeft}</a>
+    <a class="tb-toolbar-new-settings tb-icons" href="javascript:void(0)" title="toolbox settings">${icons.settings}</a>
     <label class="tb-first-run">&#060;-- Click for settings</label>
     <span id="tb-bottombar-contentleft">
         <span id="tb-toolbarshortcuts"></span>
@@ -215,13 +216,13 @@ export default new Module({
         <span id="tb-toolbarcounters">
             <a title="no mail" href="${
         TBCore.link('/message/inbox/')
-    }" class="nohavemail tb-icons" id="tb-mail">${TBui.icons.userInbox}</a>
+    }" class="nohavemail tb-icons" id="tb-mail">${icons.userInbox}</a>
             <a href="${TBCore.link('/message/inbox/')}" id="tb-mailCount"></a>
             <a href="${newModmailUrl}" class="nohavemail access-required tb-icons" id="tb-new_modmail" ${
         openMailTab ? 'target="_blank"' : ''
-    }>${TBui.icons.modmail}</a>
+    }>${icons.modmail}</a>
             <a href="${newModmailUrl}" id="tb-new-modmailcount" ${openMailTab ? 'target="_blank"' : ''}></a>
-            <a title="modqueue" href="${modQueueUrl}" id="tb-modqueue" class="tb-icons">${TBui.icons.modqueue}</a>
+            <a title="modqueue" href="${modQueueUrl}" id="tb-modqueue" class="tb-icons">${icons.modqueue}</a>
             <a href="${modQueueUrl}" id="tb-queueCount"></a>
         </span>
     </span>
@@ -291,7 +292,7 @@ export default new Module({
                 : `/r/${unmoderatedSubreddits}/about/unmoderated`,
         );
         $modBar.find('#tb-toolbarcounters').append(`
-<a title="unmoderated" href="${unModQueueUrl}" class="tb-icons" id="tb-unmoderated">${TBui.icons.unmoderated}</a>
+<a title="unmoderated" href="${unModQueueUrl}" class="tb-icons" id="tb-unmoderated">${icons.unmoderated}</a>
 <a href="${unModQueueUrl}" id="tb-unmoderatedCount"></a>
 `);
     }
@@ -299,7 +300,7 @@ export default new Module({
     const $modbarhid = $(`
 <div id="tb-bottombar-hidden" class="${compactHide ? 'tb-bottombar-compact' : ''}">
     <a class="tb-bottombar-unhide tb-icons" href="javascript:void(0)">${
-        compactHide ? TBui.icons.dotMenu : TBui.icons.arrowRight
+        compactHide ? icons.dotMenu : icons.arrowRight
     }</a>
 </div>
 `);
@@ -357,24 +358,24 @@ export default new Module({
                         <td class="tb-my-subreddits-subreddit">
                             <a title="/r/${this.subreddit} modqueue" target="_blank" href="${
                     TBCore.link(`/r/${this.subreddit}/about/modqueue`)
-                }" data-type="modqueue" data-subreddit="${this.subreddit}" class="tb-icons">${TBui.icons.modqueue}</a>
+                }" data-type="modqueue" data-subreddit="${this.subreddit}" class="tb-icons">${icons.modqueue}</a>
                             <a title="/r/${this.subreddit} unmoderated" target="_blank" href="${
                     TBCore.link(`/r/${this.subreddit}/about/unmoderated`)
-                }" data-type="unmoderated" data-subreddit="${this.subreddit}" class="tb-icons">${TBui.icons.unmoderated}</a>
+                }" data-type="unmoderated" data-subreddit="${this.subreddit}" class="tb-icons">${icons.unmoderated}</a>
                             <a title="/r/${this.subreddit} moderation log" target="_blank" href="${
                     TBCore.link(`/r/${this.subreddit}/about/log`)
-                }" data-type="modlog" data-subreddit="${this.subreddit}" class="tb-icons">${TBui.icons.modlog}</a>
+                }" data-type="modlog" data-subreddit="${this.subreddit}" class="tb-icons">${icons.modlog}</a>
                             <a title="/r/${this.subreddit} traffic stats" target="_blank" href="${
                     TBCore.link(`/r/${this.subreddit}/about/traffic`)
-                }" data-type="traffic" data-subreddit="${this.subreddit}" class="tb-icons">${TBui.icons.subTraffic}</a>
+                }" data-type="traffic" data-subreddit="${this.subreddit}" class="tb-icons">${icons.subTraffic}</a>
                             ${
                     usernotesEnabled
-                        ? `<a title="/r/${this.subreddit} usernotes" href="javascript:;" class="tb-un-config-link tb-icons" data-subreddit="${this.subreddit}">${TBui.icons.usernote}</a>`
+                        ? `<a title="/r/${this.subreddit} usernotes" href="javascript:;" class="tb-un-config-link tb-icons" data-subreddit="${this.subreddit}">${icons.usernote}</a>`
                         : ''
                 }
                             ${
                     configEnabled
-                        ? `<a title="/r/${this.subreddit} config" href="javascript:;" class="tb-config-link tb-icons" data-subreddit="${this.subreddit}">${TBui.icons.tbSubConfig}</a>`
+                        ? `<a title="/r/${this.subreddit} config" href="javascript:;" class="tb-config-link tb-icons" data-subreddit="${this.subreddit}">${icons.tbSubConfig}</a>`
                         : ''
                 }
                         </td>
@@ -463,7 +464,7 @@ export default new Module({
                     target="_blank"
                     title="this is a ${TBCore.buildType} build of toolbox. click to copy version information"
                 />
-                    <i class="tb-icons">${TBui.icons.prerelease}</i>
+                    <i class="tb-icons">${icons.prerelease}</i>
                     <span>${TBCore.toolboxVersion}</span>
                 </button>
             `).on('click', () => {
@@ -489,7 +490,7 @@ export default new Module({
     if (debugMode) {
         // Reload button
         $('#tb-bottombar').find('#tb-toolbarcounters').before(
-            `<a href="javascript:;" id="tb-reload-link" class="tb-icons" title="reload toolbox">${TBui.icons.tbReload}</a>`,
+            `<a href="javascript:;" id="tb-reload-link" class="tb-icons" title="reload toolbox">${icons.tbReload}</a>`,
         );
 
         $body.on('click', '#tb-reload-link', () => {
