+++ extension/data/modules/syntax.js
@@ -1,8 +1,7 @@
 import CodeMirror from 'codemirror';
 import $ from 'jquery';
 
-import {syntaxHighlighterThemeSelect as themeSelect} from '../tbconstants.ts';
-import {Module} from '../tbmodule.jsx';
+import {Module, syntaxHighlighterThemeSelect as themeSelect} from '../tbmodule.jsx';
 import * as TBui from '../tbui.js';
 
 export default new Module({
