+++ extension/data/util/icons.ts
@@ -1,0 +1,40 @@
+/**
+ * Material design icons mapped to toolbox names through their hexcode.
+ *
+ * @example
+ * $(`<div class="tb-icon">${icons.NAME}</div>`)
+ */
+export const icons = {
+    add: '&#xe145;', // add
+    addBox: '&#xe146;', // add_box
+    addCircle: '&#xe148;', // add_circle_outline
+    arrowLeft: '&#xe314;', // keyboard_arrow_left
+    arrowRight: '&#xe315;', // keyboard_arrow_right
+    close: '&#xe5cd;', // close
+    comments: '&#xe0b7;', // chat
+    delete: '&#xe872;', // delete
+    dotMenu: '&#xe5d4;', // more_vert
+    edit: '&#xe3c9;', // edit
+    help: '&#xe8fd;', // help_outline
+    history: '&#xe889;', // history
+    list: '&#xe896;', // list
+    modlog: '&#xe3ec;', // grid_on
+    modqueue: '&#xe8b2;', // report_problem
+    modmail: '&#xe156;', // inbox
+    overlay: '&#xe8ea;', // view_array
+    profile: '&#xe853;', // account_circle
+    refresh: '&#xe5d5;', // refresh
+    remove: '&#xe15b;', // remove
+    settings: '&#xe8b8;', // settings
+    sortDown: '&#xe5db;', // arrow_downward
+    sortUp: '&#xe5d8;', // arrow_upward
+    subTraffic: '&#xe6e1;', // show_chart
+    tbConsole: '&#xe868;', // bug_report
+    tbReload: '&#xe86a;', // cached
+    tbSettingLink: '&#xe157;', // link
+    tbSubConfig: '&#xe869;', // build
+    unmoderated: '&#xe417;', // remove_red_eye
+    userInbox: '&#xe0be;', // email
+    usernote: '&#xe06f;', // note
+    prerelease: '&#xe86f;', // code,
+};
